requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
openai==2.29.0
langchain==1.2.13
langchain-openai==1.1.11
langchain-qdrant==1.1.0
langchain-community
langchain-core
langgraph
langgraph-checkpoint
pymongo==4.15.5
qdrant-client==1.17.1
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
tenacity==9.1.4
motor==3.7.1